Fix: dto 수정

+++ back/src/main/java/com/gamul/api/response/MyRecipeIngredientInfoRes.java
@@ -31,7 +31,7 @@ public class MyRecipeIngredientInfoRes {
     boolean basket;
 
     @JsonProperty("high_class_id")
-    Long highClassId;
+    int highClassId;
 
     @JsonProperty("high_class_name")
     String highClassName;
